unify config and network resources to utils
