--- movai_core_shared/common/network.py
@@ -1,12 +1,0 @@
-import socket
-
-
-def get_ip_address() -> str:
-    """Returns the host(container) ip address.
-
-    Returns:
-        str: The ip address of the host.
-    """
-    hostname = socket.gethostname()
-    ip_addr = socket.gethostbyname(hostname)
-    return ip_addr
